workaround the fact that gmp requires the modulo to be odd.
When doing a powModSec, gmp FPE if the modulo is even.
In the case where the modulo is even, we use the standard non secure version.
Document this fact, and also note that it is fine as for RSA, DH, DSA ..
the modulo is either prime or the product of primes.